Genome IDs for watching species numbers
Will use for win/loose conditions.

Every gene now can have an enum ID assigned to it.
When a new species is spawned a it's active DNA ID is calculated.
DNA ID is an integer with bit flags set to 1 at positions that are the
same as enum ID of every gene present in the active genes list.

DNA ID is used to compare new species to our existing presets at
runtime.

In GameController component you can now set an animal preset to watch.
If you set it to giraffe, then the controller will count how many
giraffes are alive (see UpdateSpeciesWatch() method)